+++ Assets/AShooter/Scripts/User/Models/MoneyMetta/CoinMetta.cs
@@ -6,10 +6,10 @@ namespace User
     public class CoinMetta : ICoinMetta
     {
 
-        public float Value { get; private set; }
+        public int Value { get; private set; }
 
 
-        public CoinMetta(float countHealth)
+        public CoinMetta(int countHealth)
         {
             Value = countHealth;
         }
